refactor: Rename getMessageById to getMessageContentById
This is to better reflect the purpose of the function - it returns
`message.raw_content`.
